+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Changed the `Command` color config name to `Footer` for symmetry with the `Header` color config name.
 - Fixed whitespace and control code coloring in the viewer.
 - Fixed left mouse button sometimes not responding.
+- Fixed piped input.
 - Fixed "Access denied" error when using `Ctrl-Z` to undo saved hex edits.
 - Fixed `Ctrl-Z` so it also works when in hex view mode (versus edit mode).
 - Fixed `F7` and `F8` to also work in hex view mode (versus edit mode).